review: tighten module boundaries and signal a fully-blocked push
Addresses the second review round.

Module boundaries:
- `scan` no longer depends on `parse`: the index-time `Preprocessor` trait +
  `RedactSecrets` (its only impl) are gone, replaced by a plain
  `index::redact_turns(&mut turns, scanner)`. `scan` now imports nothing from
  the crate — pure detection/redaction over text and `Finding`s. Its module doc
  no longer enumerates callers.
- `scrub` moves out of `index` into its own module — it's a store-remediation
  command, not part of indexing.
- `chunk::chunks_from_batches` moves to `chunk` (it reconstructs `Chunk`s).
- `embed_batched`'s doc no longer names its callers.

Behavior:
- A push that the secret gate holds back entirely now returns `Pushed { blocked
  }`, and the CLI exits non-zero — automation can no longer mistake "nothing
  published, secrets held back" for success.

Docs:
- `scrub --help` states it cleans the local store only; it does not scrub an
  already-published remote (the push gate only stops adding to one).
